requirements.txt
fastapi
uvicorn
openai
chromadb
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
asyncio